Updates for review comments
