+++ parquet/util.py
@@ -1,4 +1,11 @@
 import os
+import pandas as pd
+
+
+class ParquetException(Exception):
+    """Generic Exception related to unexpected data format when
+     reading parquet file."""
+    pass
 
 
 def sep_from_open(opener):
@@ -18,3 +25,23 @@ def default_mkdirs(f):
 
 def default_open(f):
     return open(f, 'rb')
+
+
+def val_to_num(x):
+    # What about ast.literal_eval?
+    try:
+        return int(x)
+    except ValueError:
+        pass
+    try:
+        return float(x)
+    except ValueError:
+        pass
+    try:
+        return pd.to_datetime(x)
+    except ValueError:
+        pass
+    try:
+        return pd.to_timedelta(x)
+    except:
+        return x
